nose el py test no fufa

+++ phyton/ejercicios U1/ejercicio_2-2.py
@@ -1,4 +1,6 @@
 HT = int(input("Horas de trabajo: "))
-CH = int(input("Coste por hora: "))
+CH = float(input("Coste por hora: "))
 IT = CH * HT
 print("Importe total: " + str(IT))
+
+
